MAJOR UPDATE: juci $config interface has changed.
Now we always load /etc/config/juci. The format of the config has been updated
and /www/config.json is not used anymore. This means a single config is always
used now. Note: the /etc/config/juci needs to be readable by users who are not
logged in. Check in juci.js file to see what new sections are supported by
dafult.

This change is dependent on subsequent commits that will be pushed on top of it.