juci: add simple gui level capability code (to hide some elements if they will not work anyway)
Note: use access level controls are a mess. They currently include a lot of
legacy stuff from rpcd and this area does need quite a lot of refactoring. A
proper capability and role based access control system needs to be designed for
juci that will be used by both backend and frontend.